add P2TR key path code (#1)
This MR adds the code for constructing transactions for spending from so-called P2TR untweaked key path addresses.

This involves creating a transaction of a type different than what is implemented for P2PKH, which a pre-Segwit format i.e. not using witnesses, which are kind of the default today. But still, the serialized format is slightly different with and without witnesses (see BIP141).

Finally, the transaction is signed by constructing a hash-to-be-signed in a very specific way, which is described in BIP341.

Beyond P2TR, this MR also pretty-formats a few Motoko files, which I accidentally did and there's no easy way back, but since we want formatted code at some point, we should do this in a separate MR because very few files are formatted in the canonical Motoko format.